+++ backend/main.py
@@ -60,7 +60,7 @@ class ScreenEndpoint(ElementEndpointBase):
     _session_cls = Session
     _element = Screen
     _owner_attr = 'user_id'
-    _other_readable = list(['id', 'desc', 'template_id', 'user_id', 'duration', 'till', 'repeat', 'loop', 'variables', 'locked', 'displayed', 'key'])
+    _other_readable = list(['id', 'desc', 'template_id', 'user_id', 'duration', 'till', 'repeat', 'loop', 'variables', 'locked', 'displayed', 'default', 'key'])
     _other_createable = list(['desc', 'template_id', 'user_id', 'duration', 'till', 'repeat', 'loop', 'variables'])
     _all_readable = list(['id', 'duration', 'till', 'repeat', 'loop', 'variables', 'key'])