+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 
   * It's now possible to connect a Discord Bot to your server, that fetches the player activitiy. The counts of the played games can then be displayed on a Kiosk. For more information see: [docs/discord-playercount.md](docs/discord-playercount.md)
   * added an Prometheus endpoint, that can be enabled in Settings. This currently only provides collected data from Discord
+
+
+- Streamer Wizard
+- default Timeline on Kiosk